Add files via upload
QR code to view the contents of AR3. Point the camera at it, read the QR, point the camera at it again. Used barcode markers found here: https://laboratoryforplayfulcomputation.github.io/arcadia/docs/markers.html and put the barcode for 3 into qr-codemonkey.com. (Note that these barcodes are *not* the same thing as ARUCO markers, which I was experimenting with before.)